pyproject.toml
[tool.poetry]
name = "bayegani"
version = "0.1.0"
description = ""
authors = ["Mahdi Sadeghi <mahdi74sadeghi@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.7"
click = "^8.1.3"
pyyaml = "^6.0"
pika = "^1.3.1"
pymongo = "^4.3.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
